queue: split struct QueueStateSync into submodule
Split `struct QueueStateSync` into a submodule.
